# a7_openapi.yaml
openapi: 3.0.3
info:
  version: '1.0'
  title: 'MyGarden API'
  description: 'Web Resources Specification (A7) for MyGarden'
servers:
  - url: http://lbaw2131.lbaw-prod.fe.up.pt/
    description: Production server
...
paths:
...
  /item/{id}:
    get:
      operationId: R209
      summary: 'R209: Item Page'
      description: 'Provide the item description, images, reviews, price and allow placing order. Access: PUB'
      tags:
        - 'M02: Items'

      parameters:
        - $ref: '#/components/parameters/ID'

      responses:
        '200':
          description: 'Ok. Show [UI12](http://lbaw2131-piu.lbaw-prod.fe.up.pt/misc/product_detail.php)'
...
  /api/item:
    get:
      operationId: R622
      summary: 'R622: Item List'
      description: 'Provide a list of Items. Access: PUB'
      tags:
        - 'M02: Items'

      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Search'      # O que é isso?
        - in: query
          name: supplierId
          schema:
            type: string
          required: false

      responses:
        '200':
          description: 'Successfully returned a list of Items'
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ItemInList'

    post:
      operationId: R623
      summary: 'R623 - Item, tag e coupon'
      description: 'Creates item (description, unit, stock, seller, image). Acess: USR'
      tags:
        - 'M02: Items'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - supplier_id
                - name
                - price
                - quantityAvailable
                - description
                - images

              properties:
                supplier_id:
                  type: integer
                name:
                  type: string
                description:
                  type: string
                quantityAvailable:
                  type: integer
                price:
                  type: number
                type:
                  type: string
                  
                images:
                  type: array
                  items:
                    type: string
                    format: byte
                    description: Base64-encoded contents of image  
                
                tags:
                  type: array
                  items:
                    type: integer           # TODO: não devia ser strings?
                
                bundle:
                  type: boolean

      responses:
        '204':
          description: Successfully added/updated item
  /api/item/{id}:
    get:
      operationId: R624
      summary: 'R624 - Item, tag e coupon'
      description: 'Returns item’s information (description, unit, stock, seller, image) and its reviews and tags. Access: PUB'
      tags:
        - 'M02: Items'
      parameters:
        - $ref: '#/components/parameters/ID'

      responses:
        '200':
          description: Successfully returned item's data and reviews
          content:
            application/json:
              schema:
                # tree with all the information needed
                type: object      # TODO: Secalhar componente?
                items:
                  properties:
                    supplierID:
                      type: integer
                    description:
                      type: string
                    quantityAvailable:
                      type: integer
                    price:
                      type: number
                    rating:
                      type: number
                    bundle:
                      type: boolean

    put:
      operationId: R625
      summary: 'R625 - Item, tag e coupon'
      description: 'Updates item (description, unit, stock, seller, image) and its reviews and tags. Access: OWN'
      tags:
        - 'M02: Items'
      parameters:
        - $ref: '#/components/parameters/ID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                description:
                  type: string
                quantityAvailable:
                  type: integer
                price:
                  type: number
                bundle:
                  type: boolean

      responses:
        '204':
          description: Successfully updated item

        '404':
          $ref: '#/components/responses/ItemNotFound'

        '405':
          description: Method not allowed

    delete:
      operationId: R626
      summary: 'R626 - Item, tag e coupon'
      description: 'Removes a seller s item. Access: OWN, ADM'
      tags:
        - 'M02: Items'

      parameters:
        - $ref: '#/components/parameters/ID'


      responses:
        '204':
          description: Successfully deactivated item

        '404':
          $ref: '#/components/responses/ItemNotFound'
...
components:
  schemas:
...
    ItemInList:
      type: object
      properties:
        item_id:
          type: integer
        supplier_id:
          type: integer
        supplier_name:
          type: string
        supplier_location:
          type: string
        item_description:
          type: string
        item_price:
          type: number
        item_unit:
          type: string
        item_picture:
          type: string
          format: byte
...
  parameters:
...
    ID:
      name: id
      in: path
      required: true
      description: The ID number
      schema:
        type: integer
    Limit:
      name: limit
      in: query
      description: 'Limits the number of results to return'
      schema:
        type: integer
    Search:
      name: search
      in: query
      description: 'Filters results based on search criteria'
      schema:
        type: string
...
  responses:
    ItemNotFound:
      description: The item was not found
      content:
        application/json:
          schema:
            type: string
...